% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aibot.R
\name{ai}
\alias{ai}
\title{处理结果}
\usage{
ai(keyword, n = 3, low = 0.6, high = 0.9, detail = FALSE)
}
\arguments{
\item{keyword}{关键词}

\item{n}{返回数}

\item{low}{最小值}

\item{high}{最大值}

\item{detail}{是否显示明细}
}
\value{
返回列表
}
\description{
处理结果
}
\examples{
ai()
}
