% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aibot.R
\name{ai2}
\alias{ai2}
\title{处理结果}
\usage{
ai2(keyword, n = 3, low = 0.6, high = 0.9, detail = TRUE)
}
\arguments{
\item{keyword}{关键词}

\item{n}{返回数}

\item{low}{最小值}

\item{high}{最大值}

\item{detail}{是否显示明细}
}
\value{
返回列表
}
\description{
处理结果
}
\examples{
ai2()
}
