% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aibot.R
\name{ai2_env}
\alias{ai2_env}
\title{针对对虚拟环境的支持}
\usage{
ai2_env(keyword, n = 3, low = 0.6, high = 0.9, detail = TRUE)
}
\arguments{
\item{keyword}{关键词}

\item{n}{数量}

\item{low}{最小}

\item{high}{最大}

\item{detail}{明细}
}
\value{
返回值
}
\description{
针对对虚拟环境的支持
}
\examples{
ai2_env()
}
