% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aibot.R
\name{ai_env}
\alias{ai_env}
\title{增加对env的支持}
\usage{
ai_env(keyword, n = 3, low = 0.6, high = 0.9, detail = FALSE)
}
\arguments{
\item{keyword}{关键词}

\item{n}{数量}

\item{low}{最小值}

\item{high}{最大值}

\item{detail}{明细}
}
\value{
返回值
}
\description{
增加对env的支持
}
\examples{
ai_env()
}
