% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aibot.R
\name{aibot}
\alias{aibot}
\title{创建查询机器人}
\usage{
aibot(keyword = "发现神行多少钱", n = 3, python = "/usr/local/bin/python3")
}
\arguments{
\item{keyword}{关键词}

\item{n}{返回结果数}

\item{python}{指定python路径}
}
\value{
返回值
}
\description{
创建查询机器人
}
\examples{
bb <- aibot('发现神行多少钱')
print(bb)
}
