% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aibot.R
\name{aibot_answer_type}
\alias{aibot_answer_type}
\title{增加答案的回复类型}
\usage{
aibot_answer_type(aibot_res, min = 0.6, high = 0.9)
}
\arguments{
\item{aibot_res}{原始答案}

\item{min}{最小值}

\item{high}{最大值}
}
\value{
返回值
}
\description{
增加答案的回复类型
}
\examples{
aibot_answer_type()
}
