% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aibot.R
\name{aibot_env}
\alias{aibot_env}
\title{增加机器人查询测试env}
\usage{
aibot_env(keyword = "发现神行多少钱", n = 3)
}
\arguments{
\item{keyword}{查看词}

\item{n}{数量}
}
\value{
返回值
}
\description{
增加机器人查询测试env
}
\examples{
aibot_env()
}
